Add useLayoutState as part of DraggableCanvasWindow file
